changed bot to deal with individual Attribute pages
