+++ src/deadbot.py
@@ -1,0 +1,37 @@
+import os
+import mwclient
+from utils import json, pages
+
+'''
+DeadBot pulls all aggregated data for heros, items ,buildings and more to populate
+attribute values on the deadlock wiki via the MediaWiki API
+
+A DeadBot user has been created for this purpose, so the password will be needed
+to run this locally. However local usage should be limited exclusively to testing
+'''
+class DeadBot:
+    def __init__(self):
+        self.edit_counter = 0
+
+        site = mwclient.Site('deadlocked.wiki', path='/')
+        site.login('DeadBot', os.environ.get('BOT_WIKI_PASSWORD'))
+
+        self.attribute_pages = [page for page in site.pages if pages.page_has_category(page, 'Category:Attribute')] 
+        
+    def run(self):
+        for page in self.attribute_pages:
+            self._update_page(page)
+        
+    def _update_page(self, page):
+        updated_text = "all the stats"
+        
+        # Save the page with the updated content
+        if self.edit_counter != 0:
+            page.save(updated_text, summary="DeadBot auto-update")
+            print(f"Page '{page.name}' updated - {self.edit_counter} new changes")
+        else:
+            print(f"No changes made to '{page.name}'")
+
+if __name__ == '__main__':
+    bot = DeadBot()
+    bot.run()
