+++ src/parser/README.md
@@ -3,7 +3,6 @@ Data fetcher that builds the master_table file that is consumed by DeadBot
 
 ## Fetched Data
 Takes raw source files and exports data to more usable json formats\
-[See README for details](./data-fetcher/README.md)
 
 **Currently fetches the following**\
 Hero\
