+++ src/parser/attr-maps/hero-map.json
