+++ src/parser/localization_parser.py
